mejoras

+++ KaphiyQuipu.ViewModels/NotaIngresoAcopio/ConsultaPorIdNotaIngresoAcopioDTO.cs
@@ -32,8 +32,8 @@ namespace KaphiyQuipu.DTO
         public string Calidad { get; set; }
         public string TipoCertificacion { get; set; }
         public int TotalSacos { get; set; }
-        public int PesoSaco { get; set; }
-        public int PesoKilos { get; set; }
+        public decimal PesoSaco { get; set; }
+        public decimal PesoKilos { get; set; }
         public string ObservacionesSolicitudCompra { get; set; }
         public decimal SacosPC { get; set; }
         public decimal KilosBrutosPC { get; set; }
